requirements.txt
google-api-core==1.23.0
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
ipython==7.16.1
ipython-genutils==0.2.0
oauth2client==4.1.3
pocketsphinx2==0.1.17
PyAudio==0.2.11
SpeechRecognition==3.8.1
numpy==1.19.4
scipy==1.5.4
sounddevice==0.4.1
openai-whisper
